[adding-europe] Moved spokenCurrency to currency file and added the EUR spoken field.
